Conflicts:
	TitaniumTestTi/Resources/$variables.rjss
	TitaniumTestTi/Resources/android/appicon.png
	TitaniumTestTi/Resources/android/default.png
	TitaniumTestTi/Resources/app.js
	apidoc/Titanium/Media/AudioPlayer.yml
	iphone/Classes/TiMediaAudioPlayerProxy.m
	iphone/Classes/TiUIButtonBarProxy.m

+++ iphone/Classes/NSData+Additions.m
@@ -6,7 +6,6 @@
  */
 
 #import "NSData+Additions.h"
-#import "Base64Transcoder.h"
 #import <CommonCrypto/CommonCryptor.h>
 
 #pragma mark Hex
@@ -45,7 +44,7 @@ NSData * dataWithHexString (NSString * hexString)
 	// byte) to the NSData result
 	int i;
 	char current;
-	const int size = [hexString length] / 2;
+	const NSUInteger size = [hexString length] / 2;
 	const char * stringBuffer = [hexString cStringUsingEncoding:NSASCIIStringEncoding];
 	NSMutableData * result = [NSMutableData dataWithLength:size];
 	char * resultBuffer = (char*)[result mutableBytes];
